requirements/local.txt
-r base.txt

Werkzeug==2.0.2
ipdb==0.13.9
psycopg2==2.9.2
watchgod==0.7
pylint-celery==0.3
factory-boy==3.2.1
django-extensions==3.1.5
